backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
tensorflow>=2.15.0
numpy>=1.26.0
